requirements.txt
streamlit
langchain
langchain-openai
langchain-chroma
chromadb
pdfminer.six
httpx
python-dotenv
tiktoken
